% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_off_list_groups.R
\name{add_off_list_groups}
\alias{add_off_list_groups}
\title{Add off-list groups to generated word-lists}
\usage{
add_off_list_groups(data)
}
\arguments{
\item{data}{}
}
\value{
a dataframe with a mutated group variable
}
\description{
Provides "number", "proper-noun", "pronoun", & "off-list" groupings for words that do not appear in the word lists you are using as a reference.
This should be used after producing wordlists with \code{word.lists::get_wordlists}, \code{tidytext::unnest_tokens} etc. AND
joining it to a word list with \code{lemma} and \code{group} variables
}
