% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doc_level_analytics.R
\name{doc_level_analytics}
\alias{doc_level_analytics}
\title{get document level analytics}
\usage{
doc_level_analytics(data, text_col = text)
}
\arguments{
\item{data}{a dataframe with entire texts in one column}

\item{text_col}{the column with the texts in}
}
\value{
dataframe with added formality variable
}
\description{
for things that can't be done with \code{udpipe()} output
this is very much a WIP
}
