% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_translations.R
\name{get_definition}
\alias{get_definition}
\title{Get English definitions of individual words from nltk}
\usage{
get_definition(token, part_of_speech)
}
\arguments{
\item{token}{an English word to pass to nltk's synset search}

\item{part_of_speech}{one of "v" -verb, "a" - adjective, "n" - noun, or "r" - adverb}
}
\value{
a character vector of length one, with definitions separated by ";\n"
}
\description{
This will match up to 3 meanings where the requested word and part of speech match.
For more information on synsets, see the
Open Multilingual Wordnet documentation.
}
