% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_translations.R
\name{get_translation}
\alias{get_translation}
\title{Get translations of individual words from nltk}
\usage{
get_translation(token, part_of_speech, language = "jpn")
}
\arguments{
\item{token}{an English word to pass to nltk's synset search}

\item{part_of_speech}{one of "v" -verb, "a" - adjective, "n" - noun, or "r" - adverb}

\item{language}{check the list of options with wordlists::ntlk_languages. Defaults to japanese}
}
\value{
a character vector of length one
}
\description{
This will match up to 4 meanings where the requested word and part of speech match.
The first 2 results from each meaning-set are returned, then the first four unique
translations are selected. This means there is a fairly strong tendency towards the
most common two or three synsets. For more information on synsets, see the
Open Multilingual Wordnet documentation.
}
