% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{list_academic}
\alias{list_academic}
\title{NAWL List}
\format{
A data frame with 3807 observations and 3 variables
\describe{
\item{lemma}{base form of the word, English}
\item{group}{"difficulty grouping"}
\item{on_list}{"academic"}
}
}
\source{
\url{https://www.newgeneralservicelist.org/nawl-new-academic-word-list/}
}
\usage{
list_academic
}
\description{
A dataset containing the New Academic Word List (NAWL) & New General Service List (NGSL).
Difficulty groupings have been arbitrarily set by me, as follows:
}
\details{
\itemize{
\item Group 1: first 500 words of NGSL by frequency & "supplementary" words  - months/numbers etc
\item Group 2: next 500 words of NGSL by frequency
\item Group 3: next 1000 words of NGSL by frequency
\item Group 4: remaining NGSL words by frequency (about 800 words)
\item Group 5: academic word list (about 950 words)
}
}
\note{
LICENSE:
Creative Commons Attribution-ShareAlike 4.0 International License.
}
\keyword{datasets}
