% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{list_aswl}
\alias{list_aswl}
\title{Academic Spoken Word List}
\format{
A data frame with 6914 observations and 3 variables
\describe{
\item{lemma}{base form of the word, English}
\item{group}{"difficulty grouping"}
\item{on_list}{"flemma"}
}
Groups are separated by level as in the paper:
\itemize{
\item Group 1: 825 headwords/3618 flemmas
\item Group 2: 450 headwords/1816 flemmas
\item Group 3: 380 headwords/1305 flemmas
\item Group 4: 141 headwords/175 flemmas
}
}
\source{
\url{https://osf.io/gwk45/}

\url{https://onlinelibrary.wiley.com/doi/10.1111/lang.12253}
}
\usage{
list_aswl
}
\description{
For more information and rationale, see the source documentation.
}
\details{
A dataset containing 6914 flemmas based on headwords as listed in The Academic Spoken Word List by Dang, Coxhead & Webb.
}
\keyword{datasets}
