% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{list_dolch}
\alias{list_dolch}
\title{New Dolch List}
\format{
A data frame with 3491 observations and 3 variables
\describe{
\item{headword}{base of the word family, English}
\item{word}{word for linking to the input text}
\item{lemma}{lemma}
\item{on_list}{"dolch"}
\item{group}{empty, but necessarily extant for shiny app}
}
}
\source{
\url{https://www.newgeneralservicelist.org/new-dolch-list}
}
\usage{
list_dolch
}
\description{
A dataset containing the extended version of the New Dolch List (NDL).
For more information and rationale, see the source documentation.
}
\details{
As the Dolch List focuses specifically on the absolute most useful words,
it has not been split into difficulty groups, though it would be possible to do so.
Please make contact if that would be useful to you.

This list contains the headwords and all members of the word families included in the
"for researchers" download from the source.
}
\keyword{datasets}
