% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{list_flemma}
\alias{list_flemma}
\title{flemma List}
\format{
A data frame with 4107 observations and 3 variables
\describe{
\item{lemma}{base form of the word, English}
\item{group}{"difficulty grouping"}
\item{on_list}{"flemma"}
}
}
\source{
\url{https://www.lextutor.ca/vp/comp/bnc_info.html}
}
\usage{
list_flemma
}
\description{
A dataset containing the 5000 most frequent flemmas
("form-based-lemma" - lemmas that can have multiple meanings with the same form)
taken from Tom Cobb's lextutor site.
Groups are separated by the hundred, and this list is specifically targeted at
graded readers.
}
\note{
Accessed February 2021
}
\keyword{datasets}
