% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{list_kelly}
\alias{list_kelly}
\title{Kelly Project English Word List}
\format{
A data frame with 7549 observations and 3 variables
\describe{
\item{lemma}{base form of the word, English}
\item{group}{"difficulty grouping"}
\item{on_list}{"flemma"}
}
}
\source{
\url{http://corpus.leeds.ac.uk/serge/kelly/}
}
\usage{
list_kelly
}
\description{
For more information and rationale, see the source documentation.
}
\details{
A dataset containing 7549 tokens, graded by CEFR difficulty. Note that while these may not be lemmas, they are noted as such to maintain consistency with other \code{list_} datasets.

Groups are separated by level as in the paper:
\itemize{
\item Group 1: A1: 789 tokens
\item Group 2: A2: 1383 tokens
\item Group 3: B1: 948 tokens
\item Group 4: B2: 921 tokens
\item Group 5: C1: 1107 tokens
\item Group 6: C2: 2401 tokens
}
}
\note{
LICENSE:
Creative Commons Attribution-NonCommercial-ShareAlike 2.0 Generic license
}
\references{
\url{https://link.springer.com/article/10.1007/s10579-013-9251-2}
}
\keyword{datasets}
