% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{list_ngsl}
\alias{list_ngsl}
\title{NGSL}
\format{
A data frame with 2848 observations and 3 variables
\describe{
\item{lemma}{base form of the word, English}
\item{group}{"difficulty grouping"}
\item{on_list}{"general"}
}
}
\source{
\url{https://www.newgeneralservicelist.org/}
}
\usage{
list_ngsl
}
\description{
A dataset containing the New General Service List
#' For more information and rationale, see the source documentation.
}
\details{
Difficulty groupings have been arbitrarily set by me, as follows:
\itemize{
\item Group 1: first 500 words of NGSL by frequency & "supplementary" words  - months/numbers etc
\item Group 2: next 500 words of NGSL by frequency
\item Group 3: next 1000 words of NGSL by frequency
\item Group 4: remaining NGSL words by frequency (about 800 words)
}
}
\keyword{datasets}
