% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{list_ngsl_all}
\alias{list_ngsl_all}
\title{Full NGSL}
\format{
A data frame with 31241 observations and 4 variables
\describe{
\item{lemma}{base form of the word, English}
\item{group}{"difficulty grouping"}
\item{on_list}{general/sup/nawl/NA - not particularly useful here, but preserved for consistency}
\item{rank}{word frequency rank according to the researchers}
}
}
\source{
\url{https://www.newgeneralservicelist.org/}
}
\usage{
list_ngsl_all
}
\description{
A dataset containing the full frequency data provided by the New General Service List researchers.
#' For more information and rationale, see the source documentation.
}
\details{
Difficulty groupings have been arbitrarily set by me, as follows:
\itemize{
\item Group 1: first 500 words of NGSL by frequency & "supplementary" words  - months/numbers etc
\item Group 2: next 500 words of NGSL by frequency
\item Group 3: next 1000 words of NGSL by frequency
\item Group 4: remaining NGSL words by frequency (about 800 words) + NAWL (about 900 words)
\item Groups 5-13: frequency groupings by first significant digit of rank (9,000-9,999, 10,000-19,999, 20,000-29,999 etc)
}
}
\note{
LICENSE:
Creative Commons Attribution-ShareAlike 4.0 International License.
}
\keyword{datasets}
