% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{list_oxford}
\alias{list_oxford}
\title{oxford 5000 cefr List}
\format{
A data frame with 11283 observations and 3 variables
\describe{
\item{cefr}{CEFR level}
\item{lemma}{base form of the word, English}
\item{pos}{part of speech - not standardised to other formats (yet)}
\item{group}{"difficulty grouping"}
\item{on_list}{"oxford_am"|"oxford_br"}
}
}
\source{
\url{https://www.oxfordlearnersdictionaries.com/wordlists/oxford3000-5000}

\url{https://www.oup.com.cn/test/oxford-3000-and-5000-position-paper.pdf}
}
\usage{
list_oxford
}
\description{
A dataset containing the 5000 most frequent words for British and American English
according to Oxford University Press.
Words with multiple POS are given one row per POS.These often have different CEFR levels in this list.
Where lemma and pos have two meanings with different CEFR levels,
the lower is used.
}
\details{
Groups are separated by CEFR level:
\itemize{
\item Group 1: A1 (1061 words)
\item Group 2: A2 (989 words)
\item Group 3: B1 (906 words)
\item Group 4: B2 (1569 words)
\item Group 5: C1 (1391 words)
}
}
\keyword{datasets}
