% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{list_wiki}
\alias{list_wiki}
\title{wiki List}
\format{
A data frame with 4999 observations and 3 variables
\describe{
\item{lemma}{base form of the word, English}
\item{group}{"difficulty grouping"}
\item{on_list}{"flemma"}
}

A data frame with 4999 observations and 3 variables
\describe{
\item{wordnet}{name of the wordnet}
\item{lang}{three-character abreviation to be passed to \code{get_syns} and \code{make_wordlist}}
\item{synsets}{number of synsets present in wordnet}
\item{words}{number of words present in wordnet}
\item{senses}{number of senses present in wordnet}
\item{core}{percentage coverage of Princeton Wordnet}
\item{licence}{licence of wordnet}

}
}
\source{
\url{http://compling.hss.ntu.edu.sg/omw/}
}
\usage{
list_wiki

list_wiki
}
\description{
A dataset containing the 5000 most frequent words
according to wikipedia in 2019
Groups are separated by the thousand.

A dataset containing information on the wordnets
provided to NLTK by OMW
Pulled from the source 2021-03-02
Follow link to source for citation information and direct download of data
}
\keyword{datasets}
