% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piped_text_analytics.R
\name{piped_text_analytics}
\alias{piped_text_analytics}
\title{get basic text analytics}
\usage{
piped_text_analytics(data, group_list, off_list = TRUE)
}
\arguments{
\item{data}{output of \code{udpipe()}}

\item{group_list}{dataframe of words/lemmas with vocab groups attatched}

\item{off_list}{if true, \code{replace_off_list()} will be called on \code{data} and \code{group_list}}
}
\value{

}
\description{
use udpipe's pos and dependency parsing to get some basic analytics on texts
}
