% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piped_text_analytics.R
\name{replace_off_list}
\alias{replace_off_list}
\title{Replace common off-list words}
\usage{
replace_off_list(data, groups)
}
\arguments{
\item{data}{output of \code{udpipe()}}

\item{groups}{dataframe of words/lemmas with vocab groups attatched}
}
\value{
a dataframe of words not on the list
}
\description{
Replace common off-list words
}
