% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_translations.R
\name{get_nltk_data}
\alias{get_nltk_data}
\title{Download nltk data}
\usage{
get_nltk_data(method = "auto", conda = "auto")
}
\arguments{
\item{method}{auto - dunno, refer to reticulate documentation}

\item{conda}{auto - dunno, refer to reticulate documentation}
}
\description{
Download nltk data
}
