% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{list_wiki}
\alias{list_wiki}
\title{wiki List}
\format{
A data frame with 4999 observations and 3 variables
\describe{
\item{lemma}{base form of the word, English}
\item{group}{"difficulty grouping"}
\item{on_list}{"flemma"}
}
}
\usage{
list_wiki
}
\description{
A dataset containing the 5000 most frequent words
according to wikipedia in 2019
}
\details{
Groups are separated by the thousand.
}
\keyword{datasets}
